Wed, Nov 30, 2022, 10:58 AM +00:00

+++ filter-less/helpers.c
@@ -8,7 +8,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
     {
         for (int j = 0; j < width; j++)
         {
-            (int) round(image[i][j]) = (image[i][j].rgbtBlue + image[i][j].rgbtGreen + image[i][j].rgbtRed)/3.0;
+            round(image[i][j]) = (image[i][j].rgbtBlue + image[i][j].rgbtGreen + image[i][j].rgbtRed)/3.0;
         }
     }
     return;
